Refactoring project structure and imports to follow our convention
