+++ src/pages/landing-page/LandingPage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import Events from "./Events";
-import NavBar from "./NavBar";
-import Categories from "./Categories"
+import Events from "./events/Events" 
+import NavBar from "../../layouts/navbar/NavBar"
+import Categories from "./categories/Categories"
 
 export default function LandingPage(){
     return(
