+++ src/pages/landing-page/categories/categories.css
