+++ src/pages/landing-page/cover.png
