+++ src/pages/landing-page/events/events.css
