+++ src/pages/landing-page/events/mock-events.json
